+++ engine/src/SomeGraphics/Rendering/IndexBuffer.cpp
@@ -6,7 +6,7 @@
 namespace sg {
 
 IndexBuffer::IndexBuffer(const std::vector<uint32_t>& indices) :
-    m_count(indices.size())
+    m_count((uint32_t)indices.size())
 {
     glCreateBuffers(1, &m_renderer_id);
     glNamedBufferStorage(m_renderer_id, m_count * sizeof(uint32_t), indices.data(), 0);
